pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "polymarket-trading"
version = "0.1.0"
description = "Polymarket 5-Min BTC Paper Trading Bot"
requires-python = ">=3.11"
dependencies = [
    "py-clob-client>=0.0.1",
    "websockets>=12.0",
    "httpx>=0.27",
    "rich>=13.7",
    "pandas>=2.2",
    "python-telegram-bot>=21.0",
    "aiosqlite>=0.20",
    "python-dotenv>=1.0",

    "matplotlib>=3.8",
    "web3>=7.14.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.3",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E", "W",     # pycodestyle
    "F",           # pyflakes
    "C901",        # 순환 복잡도
    "PLR0911",     # 함수 내 return문 과다
    "PLR0912",     # 함수 내 분기(branch) 과다
    "PLR0913",     # 함수 파라미터 과다
    "PLR0915",     # 함수 내 statement 과다
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pylint]
max-returns = 6
max-branches = 10
max-args = 5
max-statements = 40

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
